// src/lib/types.ts
export type GeneralPersonData = {
  adult: boolean;
  gender: number;
  id: number;
  known_for_department: string;
  media_type?: string;
  name: string;
  popularity: number;
  profile_path: string | null;
};
// ...
export type GeneralContent = {
  adult: boolean;
  backdrop_path: string | null;
  genre_ids: number[];
  id: number;
  media_type?: string;
  original_language: string;
  overview: string;
  popularity: number;
  poster_path: string | null;
  vote_average: number;
  vote_count: number;
};

export type GeneralMovieData = GeneralContent & {
  original_title: string;
  release_date: string;
  title: string;
  video: boolean;
};

export type GeneralShowData = GeneralContent & {
  origin_country: string[];
  original_name: string;
  first_air_date: string;
  name: string;
};
// ...
export type FullPersonDetails = GeneralPersonData & {
  also_known_as: string[];
  biography: string;
  birthday: string;
  deathday: string | null;
  homepage: string | null;
  imdb_id: string;
  place_of_birth: string;
  combined_credits: {
    cast: Array<
      (GeneralMovieData | GeneralShowData) & {
        character: string;
        credit_id: string;
        [key: string]: any;
      }
    >;
    crew: Array<
      (GeneralMovieData | GeneralShowData) & {
        credit_id: string;
        department: string;
        job: string;
        [key: string]: any;
      }
    >;
  };
};
// ...
export type RefinedGeneralContentData = {
  backdrop: string;
  tmdb_id: string;
  media_type?: string;
  overview: string;
  poster: string;
  rating: number;
  release_date: number;
  title: string;
};

export type PersonWork = RefinedGeneralContentData & {
  worked_as: string;
};

export type RefinedPersonData = {
  biography: string;
  birth: number;
  credits: {
    cast: PersonWork[];
    crew: PersonWork[];
  };
  death: number | null;
  department: string;
  gender: number;
  imdb_id: string;
  links: { label: string; value: string }[];
  name: string;
  place_of_birth: string;
  place_of_death: string;
  profile: string;
  tmdb_id: string;
};